requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==1.11.29
django-autocomplete-light==3.2.10
django-ckeditor==5.4.0
django-crispy-forms==1.7.2

django-formtools==2.1
django-import-export==1.2.0
django-js-asset==2.0.0
django-reversion==3.0.4
djangorestframework==3.8.2

et-xmlfile==1.1.0
future==0.18.2
httplib2==0.9.2
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupPy==1.14
MarkupSafe==2.1.0
mistune==2.0.2
odfpy==1.4.1
openpyxl==3.0.9
Pillow==5.1.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tablib==3.2.0
uritemplate==4.1.1
urllib3==1.26.8
xadmin @ https://github.com/the5fire/django-xadmin/archive/0.6.1.tar.gz
xlrd==2.0.1
xlwt==1.3.0
